+++ common/src/main/java/com/github/phylogeny/bettercushionplacement/EntityHelper.java
@@ -5,25 +5,27 @@ import com.github.phylogeny.bettercushionplacement.registry.CommonTags;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.core.component.DataComponents;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityTypes;
 import net.minecraft.world.entity.decoration.Cushion;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.CushionItem;
+import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.entity.EntityTypeTest;
-import net.minecraft.world.phys.AABB;
-import net.minecraft.world.phys.BlockHitResult;
-import net.minecraft.world.phys.Vec3;
+import net.minecraft.world.phys.*;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Optional;
+import java.util.function.Predicate;
 
 public class EntityHelper {
     private enum Grid {
@@ -44,14 +46,65 @@ public class EntityHelper {
         }
     }
 
-    private static boolean isSneaking(UseOnContext placeContext) {
-        return Optional.ofNullable(placeContext.getPlayer())
+    @Nullable
+    public static DyedAABB getCushionSpawnBox(Player player, Level level, HitResult hitResult) {
+        ItemStack heldStack = null;
+        for (InteractionHand hand : InteractionHand.values()) {
+            ItemStack stack = player.getItemInHand(hand);
+            if (stack.is(ItemTags.CUSHIONS)) {
+                heldStack = stack;
+                break;
+            }
+        }
+        if (heldStack == null)
+            return null;
+
+        if (hitResult instanceof EntityHitResult entityHit
+                && entityHit.getEntity() instanceof Cushion cushion)
+            hitResult = getStackedCushionHitResult(player, cushion, heldStack);
+
+        if (!(hitResult instanceof BlockHitResult blockHit)
+                || blockHit.getDirection() != Direction.UP)
+            return null;
+
+        BlockPos pos = blockHit.getBlockPos();
+        Vec3 spawnVec = hitResult.getLocation();
+        BlockHitResult blockHitResult = new BlockHitResult(
+                spawnVec,
+                Direction.UP,
+                pos,
+                false
+        );
+        UseOnContext context = new UseOnContext(player, InteractionHand.MAIN_HAND, blockHitResult);
+        UseOnContext recalculatedContext = CushionItem.recalculateContextForSpecialCollisionShapes(context);
+        if (recalculatedContext.getClickedFace() != Direction.UP)
+            return null;
+
+        BlockPlaceContext placeContext = new BlockPlaceContext(recalculatedContext);
+        Vec3 entityPos = Vec3.atCenterOfWithY(
+                pos.relative(blockHit.getDirection()),
+                placeContext.getClickLocation().y
+        );
+        spawnVec = adjustPlacementPosition(entityPos, recalculatedContext);
+        AABB spawnAABB = EntityTypes.CUSHION.getSpawnAABB(spawnVec);
+        if (intersectsCushion(level, spawnAABB) || !Cushion.wouldSuriveAt(level, spawnAABB))
+            return null;
+
+        DyeColor color = heldStack.getOrDefault(
+                DataComponents.CUSHION_COLOR,
+                DyeColor.BLUE
+        );
+        return new DyedAABB(spawnAABB, color);
+    }
+
+    private static boolean isSneaking(UseOnContext context) {
+        return Optional.ofNullable(context.getPlayer())
                 .map(Player::isSecondaryUseActive)
                 .orElse(false);
     }
 
-    private static boolean isSprinting(UseOnContext placeContext) {
-        Player player = placeContext.getPlayer();
+    private static boolean isSprinting(UseOnContext context) {
+        Player player = context.getPlayer();
         if (player instanceof ServerPlayer serverPlayer)
             return serverPlayer.getLastClientInput().sprint();
         else if (player instanceof LocalPlayer localPlayer)
@@ -94,7 +147,7 @@ public class EntityHelper {
 
     @Nullable
     public static InteractionResult checkCushionIntersection(Level level, AABB spawnAABB) {
-        return level.getEntitiesOfClass(Cushion.class, spawnAABB).isEmpty()
+        return !intersectsCushion(level, spawnAABB)
                 ? null : InteractionResult.FAIL;
     }
 
@@ -113,21 +166,33 @@ public class EntityHelper {
             Entity cushion,
             InteractionHand hand
     ) {
-        ItemStack stack = player.getItemInHand(hand);
+        ItemStack heldStack = player.getItemInHand(hand);
+        BlockHitResult hitResult = getStackedCushionHitResult(player, cushion, heldStack);
+        if (hitResult == null)
+            return null;
+
+        UseOnContext context = new UseOnContext(player, hand, hitResult);
+        return heldStack.useOn(context);
+    }
+
+    @Nullable
+    private static BlockHitResult getStackedCushionHitResult(
+            Player player,
+            Entity cushion,
+            ItemStack heldStack
+    ) {
         if (!player.isSecondaryUseActive()
-                || !stack.is(ItemTags.CUSHIONS)
+                || !heldStack.is(ItemTags.CUSHIONS)
                 || cushion.isVehicle())
             return null;
 
         Vec3 newLocation = cushion.position();
-        BlockHitResult hitResult = new BlockHitResult(
+        return new BlockHitResult(
                 newLocation,
                 Direction.UP,
                 BlockPos.containing(newLocation),
                 false
         );
-        UseOnContext context = new UseOnContext(player, hand, hitResult);
-        return stack.useOn(context);
     }
 
     public static boolean bypassCollisionBlockTags(UseOnContext context) {
@@ -138,10 +203,29 @@ public class EntityHelper {
         if (!CommonGameRules.CUSHIONS_SUPPORT_EACH_OTHER.get(level))
             return false;
 
-        return level.hasEntities(
-                EntityTypeTest.forClass(Cushion.class),
+        return intersectsCushion(
+                level,
                 anchorBox,
                 cushion -> !cushion.getBoundingBox().equals(boundingBox)
         );
+    }
+
+    private static boolean intersectsCushion(
+            Level level,
+            AABB box,
+            Predicate<Cushion> filter
+    ) {
+        return level.hasEntities(
+                EntityTypeTest.forClass(Cushion.class),
+                box,
+                filter
+        );
+    }
+
+    private static boolean intersectsCushion(
+            Level level,
+            AABB box
+    ) {
+        return intersectsCushion(level, box, _ -> true);
     }
 }